refactor(web): consolidate callout components
