+++ web/src/components/design-system/Callout/Callout.tsx
@@ -2,9 +2,23 @@
 
 import { cva } from "class-variance-authority";
 import { X } from "lucide-react";
+import { useSyncExternalStore } from "react";
 
+import useLocalStorage from "@/src/components/useLocalStorage";
 import { AlertDescription } from "@/src/components/ui/alert";
 import { Button } from "@/src/components/ui/button";
+
+const DEFAULT_STORAGE_KEY = "dismissed-callouts";
+const DEFAULT_TTL_MS = 30 * 24 * 60 * 60 * 1000;
+
+type DismissedCallout = {
+  id: string;
+  dismissedAt: number;
+};
+
+const subscribe = () => () => {};
+const getClientSnapshot = () => true;
+const getServerSnapshot = () => false;
 
 const calloutVariants = cva("relative w-full rounded-lg border p-3", {
   variants: {
@@ -26,20 +40,49 @@ const contentVariants = cva("flex min-w-0 flex-1 flex-col gap-2", {
 });
 
 export type CalloutProps = {
+  id: string;
   variant: "info" | "warning";
   align: "top" | "middle";
   children: React.ReactElement;
   actions: React.ReactElement | null;
-  onDismiss: () => void;
+  ttlMs?: number;
+  onDismiss?: () => void;
 };
 
 export function Callout({
+  id,
   variant,
   align,
   children,
   actions,
+  ttlMs = DEFAULT_TTL_MS,
   onDismiss,
 }: CalloutProps) {
+  const [dismissedCallouts, setDismissedCallouts] = useLocalStorage<
+    DismissedCallout[]
+  >(`${id}-${DEFAULT_STORAGE_KEY}`, []);
+  const isHydrated = useSyncExternalStore(
+    subscribe,
+    getClientSnapshot,
+    getServerSnapshot,
+  );
+  const dismissedCallout = dismissedCallouts.find(
+    (callout) => callout.id === id,
+  );
+  const isVisible =
+    isHydrated &&
+    (!dismissedCallout || Date.now() - dismissedCallout.dismissedAt > ttlMs);
+
+  const handleDismiss = () => {
+    setDismissedCallouts((currentCallouts) => [
+      ...currentCallouts.filter((callout) => callout.id !== id),
+      { id, dismissedAt: Date.now() },
+    ]);
+    onDismiss?.();
+  };
+
+  if (!isVisible) return null;
+
   return (
     <div role="alert" className={calloutVariants({ variant })}>
       <AlertDescription className="ml-1 flex items-start gap-2">
@@ -56,7 +99,7 @@ export function Callout({
         <Button
           variant="ghost"
           size="sm"
-          onClick={onDismiss}
+          onClick={handleDismiss}
           className="text-muted-foreground hover:text-foreground h-6 w-6 shrink-0 p-0"
           aria-label="Dismiss"
         >
